feat: change the map library to Mapbox and integrate Redux Toolkit
The map display library has been changed from Leaflet to Mapbox, which allows smooth scaling of icons and addition of labels at zooms greater than 9.5, improving the interactivity and readability of the map. In addition, state management for places has been moved from the Context API to the Redux Toolkit, improving the organization and scalability of state management.